Fix missing_keys to use blank?
missing_keys is currently doing a check for empty? || blank? this causes some older code that relied used integers in to_param to fail without a clean warning and makes it hard to debug. Is there a reason not to use blank? in this case vs the less clear nil? || empty?